+++ open-weather-wrapper/src/main/java/com/openweatherweapper/models/params/Snow.java
@@ -13,12 +13,12 @@ public class Snow {
 
     @SerializedName("3h")
     @Expose
-    private int last3hVolume;
+    private Double last3hVolume;
 
     /**
      * @return Snow volume for the last 3 hours
      */
-    public int getLast3hVolume() {
+    public Double getLast3hVolume() {
         return last3hVolume;
     }
 }
